Collapse ImgPlus <-> ImagePlus functionality
There were many different classes containing methods that do similar
things, and it can be hard to discover and compare if they aren't all
in one place.
